% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{Data.batched}
\alias{Data.batched}
\title{Combine data into Batched data}
\usage{
Data.batched(x, y, t, S)
}
\arguments{
\item{x}{n by q matrix of covariates}

\item{y}{n vector of binary responses}

\item{t}{n vector of observed time-points}

\item{S}{number of batches}
}
\value{
A list of batched data.
 \tabular{ll}{
   \code{x.batch} \tab Batched covariates \cr
   \tab \cr
   \code{y.batch} \tab Batched binary outcome \cr
   \tab \cr
   \code{t.batch} \tab Batched time-points \cr
   \tab \cr
   \code{gap.len} \tab Length of the equally spaced time interval \cr
 }
}
\description{
The time domain of oberservation was first standardized in [0,1]. Then [0,1] was divided into S equally spaced intervals as described in Jiakun et al.(2021, Biometrics).
Then data fall into each interval compose a batch of data.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
