% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.filter}
\alias{DLSSM.filter}
\title{Filtering step of Kalman Filter}
\usage{
DLSSM.filter(fit, newx, newy)
}
\arguments{
\item{fit}{A object generated by last step operation, could be DLSSM(), DLSSM.init() or DLSSM.filter().}

\item{newx}{Covariates matrix of new batch data}

\item{newy}{Binary outcome of new batch data}
}
\value{
See the returned value of DLSSM().
}
\description{
When the new data becomes available, the coefficients can be efficiently
 updated by running filtering step from the original estimates over the new observations.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
