% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.predict}
\alias{DLSSM.predict}
\title{K-steps ahead prediction}
\usage{
DLSSM.predict(fit, K, newx)
}
\arguments{
\item{fit}{A object generated by function DLSSM.init() or DLSSM.filter()}

\item{K}{K-steps ahead of prediction}

\item{newx}{optionally, a covariate matrix of subjects if it is needed to predict their probabilities.}
}
\value{
\tabular{ll}{
   \code{coef.pred} \tab Matrix with row dimension K, including one-step ahead to K-steps ahead coefficients prediction. The k-th row correspond to k-step ahead prediction. \cr
   \tab \cr
   \code{coef.pred.var} \tab Array of covariance matrix corresponding to  coef.pred. \cr
   \tab \cr
   \code{prob.pred} \tab Matrix with column dimension K, including predicted probabilities of subjects if covariates are provided. The k-th column correspond to k-step ahead prediction.  \cr
 }
}
\description{
do K-steps ahead prediction of state vector and probabilities based on current fitted model
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
