% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.smooth}
\alias{DLSSM.smooth}
\title{Smoothing step in state-space model}
\usage{
DLSSM.smooth(fit, prediction, prediction.var, filter, filter.var)
}
\arguments{
\item{fit}{A object generated by function DLSSM() or DLSSM.init() or DLSSM.filter()}

\item{prediction}{Matrix of all prediction coefficients}

\item{prediction.var}{Array of all prediction covariance matrix}

\item{filter}{Matrix of all filtering coefficients}

\item{filter.var}{Array of all filtering covariance matrix}
}
\value{
Smoothed coefficients and corresponding covariance matrix.
 \tabular{ll}{
   \code{smooth} \tab Smooothed coefficients. \cr
   \tab \cr
   \code{smooth.var} \tab Array of covariance matrix corresponding to smooth. \cr
 }
}
\description{
Smoothing past estimated coefficients.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
