% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.insur}
\alias{car.insur}
\title{Dataset contains information of full comprehensive Australian automobile insurance policies between years 2004 and 2005
A dataset containing the claim and three attributes of 67,856 policies}
\format{
A data frame with 67856 rows and 4 columns
\describe{
   \item{y}{Binary vaiable with 0 denote a policy with no claim, and 1 denote a claim policy.}
   \item{gender}{gender of deriver}
   \item{age}{age of deriver}
   \item{exposure}{period from the date of insured to the investigation, with a maximum of one year}
}
}
\usage{
car.insur
}
\description{
Dataset contains information of full comprehensive Australian automobile insurance policies between years 2004 and 2005
A dataset containing the claim and three attributes of 67,856 policies
}
\examples{
data(car.insur)
}
\references{
De Jong P et al. (2008). “Generalized linear models for insurance data.” Cambridge Books.
}
\keyword{dataset}
