% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.filter}
\alias{DLSSM.filter}
\title{model updating (filtering) when new batch of data becomes avaliable}
\usage{
DLSSM.filter(fit, newdata)
}
\arguments{
\item{fit}{last fitted model}

\item{newdata}{matrix of new batch data, each row per sample}
}
\value{
\tabular{ll}{
   \code{Filter} \tab Filtering of state \cr
   \tab \cr
   \code{Filter.var} \tab  covariance of Filter\cr
 }
}
\description{
When the new data becomes available, the coefficients can be efficiently
 updated by running filtering step based on current estimates.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
