% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.init}
\alias{DLSSM.init}
\title{Smoothing parameters selection and inital model fitting}
\usage{
DLSSM.init(data.batched, S.initial, vary, autotune = TRUE, Lambda = NULL)
}
\arguments{
\item{data.batched}{A object generated by function Data.batched()}

\item{S.initial}{How many batches of data to be used to tuning smoothing parameters}

\item{vary}{A vector specify the covariates with time-varying coefficients. The remaining covariates have constant coefficients. For example, vectorvary=(1,2) will specify first two covariates as having time-varying coefficients. Intercept is always time-varying. If vary=NULL, only intercept is time-varying coefficient.}

\item{autotune}{T/F indicates whether or not the automatic tuning procedure desribed in Jiakun et al. (2021) should be applied.  Default is true.}

\item{Lambda}{specify smoothing parameters if autotune=F}
}
\value{
See the returned value of DLSSM().
}
\description{
This function is for tuning smoothing parameters using training data. The likelihood was calculated by Kalman Filter and maximized to estimate the smoothing parameters.
For the given smoothing parameters, the model coefficients can
be efficiently estimated using a Kalman filtering algorithm.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
