% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.plot}
\alias{DLSSM.plot}
\title{Plot coefficients}
\usage{
DLSSM.plot(fit)
}
\arguments{
\item{fit}{fitted object}
}
\description{
Plot smoothed coefficients in the training part and predicted coefficients in validation part, the two parts are divided by vertical dash line.
}
\details{
If argument "fit" is an initial fitted model then only smoothed coefficients part are plotted.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
