% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.plot}
\alias{DLSSM.plot}
\title{plot coefficients}
\usage{
DLSSM.plot(fit)
}
\arguments{
\item{fit}{object generated from DLSSM() or DLSSM.init()}
}
\description{
Plot smoothed coefficients in the trainning part and predicted coefficients in validation part
}
\details{
If argument fit is initial fitted model then only smoothed coefficients part are plotted.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
