% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.plot}
\alias{DLSSM.plot}
\title{Dynamic logistic state-space prediction model for binary outcomes}
\usage{
DLSSM.plot(fit)
}
\arguments{
\item{fit}{fitted state-space model}
}
\description{
Plot predicted, filtered and smoothed coefficients
}
\details{
The argument fit could be object of DLSSM or DLSSM.init
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
