% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSSM.R
\name{DLSSM.predict}
\alias{DLSSM.predict}
\title{Prediction step of Kalman Filter}
\usage{
DLSSM.predict(fit, newx, K)
}
\arguments{
\item{fit}{A object generated by function DLSSM() or DLSSM.init() or DLSSM.filter()}

\item{newx}{Covariates matrix of subjects if you are interested in predicting their event probabilities in the future.}

\item{K}{K-steps ahead of prediction}
}
\value{
List of predicted coefficients and probabilities if covariates are provided.
 \tabular{ll}{
   \code{coef.pred} \tab Matrix with row dimension K, including one-step ahead to K-steps ahead coefficients prediction. The k-th row correspond to k-step ahead prediction. \cr
   \tab \cr
   \code{coef.pred.var} \tab Array of covariance matrix corresponding to  coef.pred. \cr
   \tab \cr
   \code{prob.pred} \tab Matrix with column dimension K, including predicted probabilities of subjects if covariates are provided. The k-th column correspond to k-step ahead prediction.  \cr
 }
}
\description{
Given the estimates of the smoothing parameters, the K-step-ahead prediction can be done by running the Kalman filtering prediction steps without the filtering steps.
}
\author{
Jiakun Jiang, Wei Yang, Stephen E. Kimmel and Wensheng Guo
}
